Cleanup auditor just a tad
